unsigned int to long in Character mac compiler error

+++ Aster/Public/TextRenderer.h
@@ -15,7 +15,7 @@ struct Character {
 	unsigned int TextureID; // ID handle of the glyph texture
 	glm::ivec2   Size;      // size of glyph
 	glm::ivec2   Bearing;   // offset from baseline to left/top of glyph
-	unsigned int Advance;   // horizontal offset to advance to next glyph
+	long Advance;   // horizontal offset to advance to next glyph
 };
